requirements.txt
fastapi
uvicorn
youtube-transcript-api
google-generativeai
python-dotenv
opencv-python
numpy
yt-dlp
ffmpeg-python
pillow
scikit-learn